3.0.6 -- New major feature! Ghost mode!
Allows you to hide the URL to the emojis by using a ghost message! This effectively lowers the character limit by 1000, but it also makes emojis in messages MUCH better!
Off by default. It's currently labeled as Experimental but also it shouldn't be buggy so if you find a bug, feel free to report it!