4.3.0 -- Nitro Activity Bypass & Custom Activities!
Major feature release:
+Developed and added a bypass to allow opening and playing Nitro Activities!
(Clients must also be running the plugin for them to be able to join!)
+Added Custom Activities!
+Added the ability to join activities that have been opened by the bypass
(Open the activity menu, click "Enable join mode", first click the Activity that the host is running (Such as, Checkers in the Park), then click Join Activity on the Watch Together activity)
Closes #35

+Fixed a bug that caused the Resolution Quick Swapper to close before you could use it.
+Replaced some uses of getBulk with getModule (cleaner code)

This update has been in the making for a while! I hope someone gets some use out of it, and happy playing!
If you find any bugs, make sure to create an issue and describe the bug in detail!