4.5.0 --  Video Codecs, Activities Removed
I regret to inform everyone that after going through the effort of fixing the Activity scripts, almost all Activities have been given a deeper level of authentication that cannot be bypassed. Furthermore, Discord has added a custom Activity URL feature themselves if you enable Developer Activity Shelf in settings. That is why, with a heavy heart, I must announce that the Activity Bypass has been removed, most likely permanently. The fixed code has been left in the plugin (although commented out) for people who are curious, and will be deleted in the next update.
The changelog is as follows:
-Activity Bypasses have been removed.
+Added Preferred Video Codec. Choose your preferred video codec between H.264, AV1, VP8, and VP9. Suggested by @userMacieG in #95
+Fixed a problem where bitrates were not applying correctly
+Fixed warnings from module filtering
+Miscellaneous small changes