5.2.8 -- Avatar Decorations Rewrite, Twitch Emotes
+Completely reworked Avatar Decorations. As a result of the system completely changing,
*****YOU WILL NEED TO RE-APPLY YOUR DECORATION!!!*****
+Fixes an issue where Avatar Decorations were completely fucking inconsistent between different clients depending on when a user happened to download the plugin.
The reason this was a little to complicated to explain, but it was dumb as fuck.
Unfortunately, due to weird Discord shenanigans, for the time being you will only be able to apply decorations that are currently in the shop.
You would think that every decoration in the shop = every decoration, but it's not; some decorations exist that aren't in the shop. Thx Discord.

+Finally found a way to detect if an emote is from Twitch. Apparently the "emoji.managed" property means that an emoji is managed by Twitch/a third party.
Knowing Discord, though, this could potentially cause another issue, but oh well. Closes #102.